Files updated from refactoring
