+++ server/base/src/main/java/org/apache/accumulo/server/cluster/RemoteShell.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.minicluster.cluster;
+package org.apache.accumulo.server.cluster;
 
 import static java.util.Objects.requireNonNull;
