+++ server/base/src/test/java/org/apache/accumulo/server/cluster/standalone/StandaloneClusterControlTest.java
@@ -16,11 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.cluster.standalone;
+package org.apache.accumulo.server.cluster.standalone;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.apache.accumulo.minicluster.cluster.standalone.StandaloneClusterControl;
 import org.junit.jupiter.api.Test;
 
 public class StandaloneClusterControlTest {
