+++ test/src/main/java/org/apache/accumulo/test/functional/LocalityGroupIT.java
@@ -37,7 +37,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
 
-import org.apache.accumulo.minicluster.cluster.standalone.StandaloneAccumuloCluster;
+import org.apache.accumulo.server.cluster.standalone.StandaloneAccumuloCluster;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.client.BatchScanner;
