+++ test/src/main/java/org/apache/accumulo/test/functional/PerTableCryptoIT.java
@@ -35,7 +35,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.apache.accumulo.minicluster.cluster.standalone.StandaloneAccumuloCluster;
+import org.apache.accumulo.server.cluster.standalone.StandaloneAccumuloCluster;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.client.admin.CloneConfiguration;
