+++ test/src/main/java/org/apache/accumulo/test/functional/ReadWriteIT.java
@@ -38,8 +38,8 @@ import javax.net.ssl.SSLSession;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
 
-import org.apache.accumulo.minicluster.cluster.ClusterControl;
-import org.apache.accumulo.minicluster.cluster.standalone.StandaloneAccumuloCluster;
+import org.apache.accumulo.server.cluster.ClusterControl;
+import org.apache.accumulo.server.cluster.standalone.StandaloneAccumuloCluster;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.clientImpl.ClientContext;
