+++ test/src/main/java/org/apache/accumulo/test/harness/TestingKdc.java
@@ -31,7 +31,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 
-import org.apache.accumulo.minicluster.cluster.ClusterUser;
+import org.apache.accumulo.server.cluster.ClusterUser;
 import org.apache.hadoop.minikdc.MiniKdc;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
